e2e tests to ensure selected region is correct after moving between certain tools.
When using Brush then Eraser or Pan, existing region should stay selected and have new additions be added to it. Similar for the Magic Wand.

We also have to ensure that the unselectRegionOnToolChange property for the Brush tool is updated on `afterCreate` so that e2e tests can dynamically update a feature flag and have it be applied.