fix: LEAP-248: Fix app crashing by multiple hotkeys
The 'unbind' function in Hotkey.ts has been enhanced to handle multiple keys provided as a string. A new helper function 'getKeys' was added to split the key string and return an array of keys. Each key in the array is then processed individually in the 'unbind' function. This helps keymaster to handle more that one hotkey.