docs: LEAP-615: Mention spaces saving for Text tag
Text tag is displayed as html, so all spaces are smashed by browser, and they also go like this to result. To avoid this users can add custom style in config, described in docs. But it only says that this affects displaying, not saving the result. Adding this note as well.