Stop reversing annotations list
Reversing the list caused problems when task is reloaded and list is reversed again.
The visual order is defined in AnnotationsCarousel, so technical order is not that important.
View All used technical order which can cause discomfort when you see different order
in different places. So that was also fixed.
Everything is behind the same FF, without it the order is an old one, reversed.